isalpha and isdigit
Write a function that receives a character and returns a positive valued integer if the character is vowel,
returns a negative valued integer if the character is consonant,
returns zero in the character is not a letter.
Use the function in a program that receives character from user and tells if the character is a vowel,
or a consonant or not a letter until the last character entered is not a letter.
isalpha() =Tests if the character is an alphabetical character. Returns '0' if it is not an alphabetic character.
isdigit() =Tests if the character is a number.
If the received character is not a number, it returns '0'.